PP-308 organise test code directory
 - renamed test/util dir to test/test_helpers to avoid confusing with app/utils
 - moved unit tests to test/unit directory

with @mcgoooo
