+++ test/test_helpers/test_renderer.js
