+++ test/transaction_details_ft_tests.js
@@ -2,7 +2,7 @@ var request = require('supertest');
 var portfinder = require('portfinder');
 var nock = require('nock');
 var app = require(__dirname + '/../server.js').getApp;
-var auth_cookie = require(__dirname + '/utils/login-session.js');
+var auth_cookie = require(__dirname + '/test_helpers/login_session.js');
 var winston = require('winston');
 
 portfinder.getPort(function (err, connectorPort) {
