+++ test/transaction_list_ft_tests.js
@@ -5,7 +5,7 @@ var portfinder = require('portfinder');
 var nock = require('nock');
 var app = require(__dirname + '/../server.js').getApp;
 var dates = require('../app/utils/dates.js');
-var auth_cookie = require(__dirname + '/utils/login-session.js');
+var auth_cookie = require(__dirname + '/test_helpers/login_session.js');
 
 var winston = require('winston');
