+++ test/transaction_search_ui_tests.js
@@ -1,7 +1,7 @@
-require(__dirname + '/utils/html_assertions.js');
+require(__dirname + '/test_helpers/html_assertions.js');
 var should = require('chai').should();
 
-var renderTemplate = require(__dirname + '/utils/test_renderer.js').render;
+var renderTemplate = require(__dirname + '/test_helpers/test_renderer.js').render;
 
 describe('The transaction list view', function () {
     it('should render all transactions', function () {
